requirements.txt
rich>=13.0.0
aiohttp>=3.9.0
dnspython>=2.6.0